fix icon behavior #701
